Flow uses new JSON-based handshake to monitor
Summary:
We want to support some additional scenarios:

(1) If a version of flow client connects that's different from the version of flow server, currently running, then currently it just terminates the flow server, but we'd like instead for the flow server to remain running and the flow client should exec the appropriate version of the flow client.

(2) If a version of flow server is running, and FlowLSP client is a different version and wants to do a once-a-second connection attempt to see if it can connect, then this current causes the existing flow server to terminate, but we'd like instead for the flow server to remain running.

But it's hard to change the client<->monitor handshake in the necessary ways...

Prior to this diff, the flow client sends a binary structure to the flow server when it does the initial handshake:
```
client_build_id: string;
client_type: ...
```

We'd like to add new fields or new ADTs to the handshake to support the new behaviors. But doing this isn't: if the server is running the prior build of flow without the new fields, and a new client sends over a handshake with additional fields, is it safe to unmarshal? (no); and if the server is running the new build of flow with the new fields, and a new client sends over a handshake without the additional fields, is it safe to unmarshal? (no).

So instead I'm moving over to a JSON string for exchanging that data which must be exchanged prior to knowing that server and client are identical versions. And then a subsequent regular ocaml marshalled structure for the other data that can be exchanged after we know they're identical (so we don't have to encode all of our ocaml world into json). Because it's JSON, we'll be able to add fields at will, and code explicitly how server and client should behave when expected fields are absent.

NOTE: This diff merely introduces the new handshake protocol. The second next diff in the stack modifies the protocol slightly to maintain backwards-compatibility between servers and clients that use old vs new handshakes.

Reviewed By: gabelevi

Differential Revision: D7990902

fbshipit-source-id: 96e9528c57dd2dd1d4fe6e8e21b74c203c889721