Flow new handshake should be backwards-compatible
Summary:
The previous diff introduced a new JSON handshake between client and monitor. But we have to address back-compat scenarios:

* A server is running which uses the old handshake, and a new client connects
* A server is running which supports the new handshake, and an old client connects

CURRENT VERSION OF THE DIFF:

The new client only ever sends on a "socks3" socket. The new server listens on the older "socks+sock2" socket and exits in response to a message on either of them; but it does the proper JSON handshake in response to sock3 messages.

This is a simple clean approach. However:

(1) Old server + new client means the client will see the lock file, but will eror "Server_socket_missing" when it tries to connect. Ephemeral clients will try to kill the server and then restart; LSP client's behavior is currently wrong but will be fixed in D8009660

(2) New server + old client means the server will immediately terminate. Old clients know how to recover in this case.

A PRIOR VERSION OF THIS DIFF:

This diff solves the problem by making so the server accepts BOTH old and new handshakes; if it receives a new-handshake-request then it replies with a new-handshake-response; if it receives an old-handshake-request then it replies with an old-handshake-response.

The plan is that for now the client will always send the old handshake. In a month or so, we can trust that all servers out there in the wild are indeed able to receive both old and new handshakes, and at that time we can deploy a client which ONLY sends the new handshake request.

There's one further detail. The way we've done LSP, its handshake contains things that aren't expressible in the old handshake. So an LSP client will always use the new handshake. The way we've encoded the new handshake, this means that an old server will always interpret an LSP connection attempt as a build_id_mismatch, and so the old server will quit.

Reviewed By: gabelevi

Differential Revision: D7995830

fbshipit-source-id: 50d056231d61431ba357d2d1843200fec890dc5c