Flow-lsp errors trim whitespace
Summary:
Think of a flow error like "'H' [1] is not an instance type" which starts with a reference to a Code.

`Errors.Friendly.message_of_group` actually produces a message string with a leading whitespace " 'H' [1] is not an instance type".

It would have been nice if I dug into the code in Errors.Friendly to discover why it's doing that, and fixed it. But I'm going to go the faster route, and just trim whitespace from the error message.

Note: previously the newtests function `doesMessageMatch` was ignoring whitespace when testing the content of an LSP message. I'd written it that way because I didn't want the whitespace generated by `JSON.stringify` to mess things up. But it meant that `doesMessageMatch` was unable to test this "whitespace in error message" issue. So I've changed `doesMessageMatch` to now pay attention to whitespace.

Note: I've replaced the old 'witherrors2.js' with a new one. I think the old one was just an idle exercise; it didn't really exercise any substantially different publishDiagnostics codepath from what 'witherrors1.js' already does.

Reviewed By: gabelevi

Differential Revision: D7862725

fbshipit-source-id: 30c697d67e741863bb3658ba336e52ea0cadb5ea