Handle in-progress diagnostic updates
Summary:
While a recheck is in progress, between StartRecheck and EndRecheck, the server will periodically send `Errors {errors,warnings}` messages. These are additive to the errors which have previously been reported. Once the recheck has finished then the server will send a new exhaustive set of errors. At this opportunity we should erase all errors not in this set.

This differs considerably from the semantics of LSP publishDiagnostics which says "whenever you send publishDiagnostics for a file, that now contains the complete truth for that file."

Prior to this diff: it acted as if every Errors message from the server had LSP semantics. This meant: (1) during incremental updates it was removing the existing errors from the file while at the same time adding the additive ones; (2) when you fixed the last error in a file, it never removed it.

After this diff: we fixed the issue. The solution has to involve keeping a record of which diagnostics have been sent previously, so we can (1) compute the result of adding diagnostics to a file, (2) compute which files no longer have errors.

DESIGN CONSIDERATIONS

The question is, where do we store the record of which diagnostics have been sent previously?

The ideal would be to store this information in the server. Indeed, the server ALREADY stores this information in its ServerEnv.errors+warnings`` fields. That way, the server could entirely avoid sending StartRecheck/EndRecheck/Errors to lspCommand, and could instead accomplish everything solely by sending LSP messages to lspCommand.

But for the moment I thought that approach was too heavyweight: the server already has a whole bunch of logic to go from ServerEnv.errors into its separate "additive/replacement" messages. For the server to support that, and also in addition the new style right next to it, led to code that was too ugly.

So I think the best solution is to put the code into lspCommand for now, so the server has a single well-defined behavior for errors. Later on once we remove the old form of persistent connection, then we can move the server over to the LSP behavior.

Reviewed By: nmote

Differential Revision: D7896579

fbshipit-source-id: e737e8da861d7ca68d7165011ca7ccfd0c230d7b