LSP shows server status during typecheck
Summary:
Flow's monitor reports its status. When Flow is idle, this status alternates between "free" and "garbage collecting". I think that the most user-relevant status reports are to show the progress of typechecking.

Prior to this diff, only the ephemeral connection reported its status (it did that every 0.5 seconds). This diff changes it so also the persistent connection reports its status.

This diff uses that report to show an LSP progress message.

Ideally I wanted to implement all of this in the flow server, rather than burdening lspCommand with extra flow-specific code. But that was awkward in this case. (because the status messages actually come from the monitor not the server and so would have to be distinguished). So I went for the more expeditious approach.

Reviewed By: gabelevi

Differential Revision: D7921953

fbshipit-source-id: 10864dfcf76fe2eccc9e70308e7661bd02f8b106