LSP should handle didSave
Summary:
Prior to this diff, flow was asking the client to send didSave requests, but was failing to handle them - producing an error in the logs.

After this diff, flow asks the client NOT to send didSave requests. But I've added a no-op handler for them in commandHandler just in case (because Nuclide has a bug where it sends didSave even if you didn't ask for it).

Reviewed By: gabelevi

Differential Revision: D7923703

fbshipit-source-id: 3e1b223936519b095a8a4c0013ff3f437121d7c4