LSP shouldn't redisplay "connecting" dialog
Summary:
If the Flow server is down, FlowLSP tries once a second to connect to the Flow server, just in case the user has opened it. It displays the result of this attempt.

If it displayed "Flow server has stopped" and the user dismisses this dialog, then we shouldn't pop up another instance of the dialog on our next attempt a second later.

lspCommand already does that correctly.

But what about if it displayed "Connecting to Flow server..." and the user dismisses this dialog? then likewise we shouldn't pop up another instance of this dialog.

This diff applies to the same code as was used in the "stopped" case, to also apply to the "connecting" case.

Reviewed By: gabelevi

Differential Revision: D7923939

fbshipit-source-id: b256ed4b4ad5aea44e67d7b34cfc512e3f413d5b