LSP init should fail if wrong version
Summary:
The various flow commands all start with an initial check in CommandUtils.assert_version - if the [version] in the .flowconfig doesn't match the current binary, then the command exits with a message "Wrong version of flow...".

This diff adds the same functionality to `flow lsp`. The right time to do the check here is in response to the "initialize" message from the client. The LSP standard is that if you can't initialize successfully, then you respond with an error, and in the error you can tell the client whether to retry or not, and if not then the client knows to close the connection. That's precisely what we do here.

Reviewed By: gabelevi

Differential Revision: D7932746

fbshipit-source-id: b90a4478cc1ce39f2309963fb76001c7725b63f5