LSP connection poll shouldn't cause server to exit
Summary:
Problem scenario: I'm working on some project that doesn't have a [version] field in its .flowconfig. There's one binary version of flow bundled within my VSCode flow-language-server, and a different version on my command-line. I've been using VSCode fine. Then at the command-line I invoke 'flow', and it (1) leaves the VSCode connection broken, (2) it doesn't display any results.

(This problem wouldn't arise if there were a [version] field in .flowconfig that unambiguously specifies which binary to launch. It only arises when two different binaries are legitimately able to be running.)

The problem has to do with the way the `flow lsp` and `flow ide` adapters both work. They poll once a second to see if the flow server has been restarted by any means (either by the user manually at the command line, or in the case of LSP by the user clicking the [Restart] button). In `flow ide` this poll happens all the time as part of status reporting; in `flow lsp` it's done only if the connection has been lost. But this poll is actually a connection attempt which carries a version field. And so if there happened to be a different binary version of the flow server running, then merely polling is enough to make the server terminate with a Build_id_mismatch error.

The solution is to use a recently-added parameter in the SocketHandshake, where we can say "Dear server, if you encounter a build id mismatch, please just hang up rather than terminating".

Note: this change won't help in the case where the server is old enough that it doesn't yet had the recently-added parameter in the SocketHandshake. Prior to that, the server always unavoidably terminated in response to build id mismatch.

HERE'S THE EXACT BLOW-BY-BLOW ACCOUNT OF HOW THINGS WENT WRONG PRIOR TO THIS DIFF.

1. Nuclide has launched "flow_v2 lsp". Right off the bat it's happy to autostart a flow server, but only once.

2. LSP tries to connect but encounters server_missing. So it uses up its once autostart by starting "flow_v2 server"

3. LSP tries to connect, and succeeds!

4. At command-line user invokes "flow_v1". It attempts to connect to the flow server.
Flow server has a version mismatch, so it exited. The command relaunches a new flow_v1 server

5. LSP attempts to connect. It gets back Server_busy(Not_responding). I guess the server is busy starting up from step 4.

6. LSP attempts to connect. It gets back Build_id_mismatch. The server exits.

7. LSP attempts to connect. It gets back Server_missing. It has already autostarted previously, and so won't do it again. It reports to the user that the flow server is stopped.

8. At command-line maybe user again invokes "flow_v1". It attempts to connect to the flow server but gets server_missing. So it launches a new "flow server v1".

9. LSP attempts to connect once a second. It gets back Server_busy(Not_responding), because the server is starting up.

10. LSP attempts to connect once a second. It gets back a Build_id_mismatch.
The server exits.

11. LSP attempts to connect. It gets back Server_missing. It has already autostarted previously, and so won't do it again. It tells the user that the flow server is stopped.

Reviewed By: gabelevi

Differential Revision: D8009660

fbshipit-source-id: fc3028b7ab19cae6d18e9ae6ccd8c5aa2a42eb67