[hack] Wrap ALL MultiWorker.call invocations in with_exit_on_exception
Summary:
There is no point in pretending that any exceptions happening there are recoverable. This is our main "log and exit" function, and I'm applying it everywhere.

Not doing it results in spooky action at a distance bugs, where a recheck job fails with "Worker_busy", that turns out to be a previous ServerInferTypeBatch call throwing an exception and attempting to recover from it. It's better to fail ASAP to easily correlate problems with reasons.

Reviewed By: gregory-nisbet

Differential Revision: D8274801

fbshipit-source-id: fb1af399bbd89ef7b61a29ddd21c2f63aa507d25