Disallow destructuring into type portions at require
Summary:
Before, an exported type `T` like the following:
```
// File A.js
export type T = number;
export default {};
```
Could be imported and destructured into a local variable:
```
// File B.js
const {T} = require('./A');
```
This is not a useful pattern since `T` cannot be used as a value in `B.js`. Sure enough
any potential use would result in an error further down the line, but it's good practice to
hint to the actual source of the error which is the import itself.

This commit, flags an error at the point of destructuring of the imports from `A.js`.

Reviewed By: avikchaudhuri

Differential Revision: D8345836

fbshipit-source-id: c1529fd4815d0859ca875aca25669a79c92083c4