LSP supports rage
Summary:
When the user clicks the bug button in Nuclide, it sends "telemetry/rage" request to the LSP server. (This is a Facebook-specific extension to the LSP protocol, but the VSCode team indicated they were interested in adding something similar to VSCode).

We've used "telemetry/rage" in Hack to gather some useful debugging information. This diff does the same for Flow.

Why use a telemetry/rage command rather than just writing stuff into logs? - it's because we don't have to be pessimistic. A log is written always all the time so you don't want to put too much stuff into it. But a telemetry/rage requiest is issued precisely when the user has observed something wrong, and is content to wait for a minute while detailed logs are gathered, and so we feel more free to gather complete data.

Reviewed By: gabelevi

Differential Revision: D8239030

fbshipit-source-id: c1665f14211c88532b76cc4f60c26363753f2fe1