LSP skips typeCoverage for non-@flow files
Summary:
What should type coverage report for a file without the flow directive? Without the directive, none of the file is being checked, so the answer should be "0% coverage" by definition.

What spans should we say are not covered by type checking? This is a weird question. Simply, the entire file is uncovered.

Prior to this diff, in a file without flow, we were indeed reporting "0%", but we were also reporting every individual expression span as being uncovered. That was confusing to read and not very actionable. Also in a large file, we ended up sending many thousands of spans sometimes in a 2MB+ json blob, and most IDEs can't cope with that many squiggles or errors in their diagnostic pane (Nuclide would just freeze up for many seconds at a time).

With this diff we give a more straightforward answer: just one single message instructing the user to use the flow directive if they want to see uncovered spans.

(Won't they still get many thousands of uncovered spans once they add the flow directive? Well, maybe, but at least they've explicitly declared their intent to work through this and get their file flow-clean).

Note: I chose to report this single sentinel message "please use the flow directive" as spanning the range (line 0 column 0)..(line 1 column 0). This displays well in Nuclide - if there's anything on the first line of the file then it will squiggle it, and it'll also have a gutter icon for the first line, and also appear in the diagnostics panel. It also works well in a completely empty file. I haven't tested in VSCode, and indeed VSCode doesn't support type-coverage. But for what it's worth VSCode is permissive in other areas of diagnostic ranges so I'm sure it will cope nicely too. https://github.com/Microsoft/language-server-protocol/issues/257

Reviewed By: panagosg7

Differential Revision: D8249855

fbshipit-source-id: 45836cb38c01e3561b8422dcb97b4ca2468ab425