LSP reports progress numbers
Summary:
Prior to this diff, lspCommand was reporting progress information like this, using a custom fb-specific extension to LSP called "window/showStatus":
```
2018-06-05 13:32:49.947 TRACE (pid:2973751) flow-language-server - LSP.trace: Received request 'window/showStatus - (18)'. Params: {
    "type": 2,
    "message": "Flow: Server is rechecking (�  merged files 0/1 (0.0%))",
    "actions": [],
}
```

But we the Nuclide UI to be able to show a percentage progress bar. We can't expect Nuclide to scrape numbers out of the human-readable text message. So this diff adds progress numbers to the status message:
```
2018-06-05 13:32:49.947 TRACE (pid:2973751) flow-language-server - LSP.trace: Received request 'window/showStatus - (18)'. Params: {
    "type": 2,
    "message": "Flow: Server is rechecking (�  merged files 0/1 (0.0%))",
    "actions": [],
    "progress": 0,
    "total": 1
}
```

The meaning is: if "progress" is present then that shows how far along we've come; if "total" is also present then that shows the total.

Reviewed By: gabelevi

Differential Revision: D8287588

fbshipit-source-id: 75796c4cf3b115c2067e641f83b2facd9f8531eb