[hack] LSP progress wire-format tweaks
Summary:
kcaze and I are converging on the message format...
```
// BEFORE THIS DIFF:
type ShowStatusParams = {
  type: number,
  actions?: Array<MessageActionItem>,
  message: string,
  shortMessage?: string,
  progress?: number,
  total?: number,
}

// AFTER THIS DIFF:
type ShowStatusParams = {
  type: number,
  actions?: Array<MessageActionItem>,
  message?: string,
  shortMessage?: string,
  progress?: {numerator: number, denominator?: number},
}
```

(1) The message is now optional. Tolstoy wrote "happy families are all alike; every unhappy family is unhappy in its own way". That's why a *green* status update doesn't even need any message.

(2) The `progress` is now bundled up into something nicer. This now means it's impossible to have a denominator without also a numerator.

Discussion: it would have been possible to alter the `ShowStatus.params` structure further in lsp.ml to encode more invariants. But the wire format is the important thing, and we're on a deadline, so I want to land this diff promptly. We can leave cleaning up lsp.ml to a future diff.

Reviewed By: kcaze

Differential Revision: D8374788

fbshipit-source-id: b140b7f53e35114bb2dd0d2ccea959a97a774f18