Make go-to-declaration smarter when used on a derived constructor.
Summary:
When a user selects a constructor call from a derived class, such as `new Derived()`, and then select "go to declaration" in their editor, we return the position of the constructor, even if `Derived` doesn't have its own explicit constructor.

Instead, we should only return the constructor if `Derived` has an explicit constructor, and if it doesn't we should return the `Derived` class itself.

Reviewed By: ljw1004

Differential Revision: D8307747

fbshipit-source-id: f0a7336da191e2cb6e58d0e1fba63884bff944d3