[PR] add RegExp.prototype.match() return type
Summary:
"If the match succeeds, the exec() method returns an array and updates properties of the regular expression object. The returned array has the matched text as the first item, and then one item for each capturing parenthesis that matched containing the text that was captured. If the match fails, the exec() method returns null."
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/exec
Closes https://github.com/facebook/flow/pull/6374

Reviewed By: gabelevi

Differential Revision: D8285744

Pulled By: mrkev

fbshipit-source-id: e7dc38bc4c8bedc2d92f78569b24364aae45ea6d