[parser] match multiple characters at once
Summary:
when lexing strings, comments, identifiers, template literals, regexes, etc, we look for a few special characters (quotes, escape codes, etc) but mostly hit the `any` case.

`any` matches a single char, so we are usually appending lexemes of a single char at a time to a buffer. we save a lot of allocations by matching longer strings.

Reviewed By: samwgoldman

Differential Revision: D8383029

fbshipit-source-id: c40056b08d720313da5eb76e16eddfa56c084b57