stricter Poly ~> Poly checking
Summary:
Currently we allow polymorphic functions to be subtypes of less polymorphic functions. This takes us
down a pretty slow path. This is especially bad when checking for compatibility between overloaded
inherited methods. Our tests don't seem to rely on such checks succeeding, so it's reasonable to try
getting rid of this flexibility.

Reviewed By: panagosg7

Differential Revision: D8102358

fbshipit-source-id: df10b55328f28795971b120fbe78e63a56142444