[hack] Multiple root support for watchman
Summary:
Hack servers only watch a single directory (the root directory). Flow servers need to watch multiple directories. This attempts to adapt the current watchman support.

The basic idea is that each watched directory needs its own

* subscription
* relative_path (from the watch root)
* clockspec

We tell watchman to watch each watched directory and subscribe to that directory. And when we run queries, we send a query for each watched directory and merge the results.

Reviewed By: avikchaudhuri

Differential Revision: D8209048

fbshipit-source-id: 4ac81f040497938ab12c22f2817eff6610d1c09a