Optimize Files.is_valid_path
Summary:
This was likely a waste of time.

I started because I was annoyed that we were doing a linear scan across file extensions. It seemed super wasteful. "Why not just look up the file extension in a set?!" wondered Gabe.

Well, it's because we allow multiple part file extensions. "foo.bar.baz" might match ".baz" or ".bar.baz".

Anyway, this might be a little better, especially in the common case of files with only a single extension.

I also did a quick pass over the callers to make sure they are properly using the currying.

Reviewed By: avikchaudhuri

Differential Revision: D8281653

fbshipit-source-id: 6ccafe378769436ddee811b753ec7744fc3f0e3b