[hack] Don't send empty title
Summary:
In LSP, optional fields generally have the form `title?: string` meaning either you omit the property, or you include it and it's a valid string.

Prior to this diff we were sending `title: ?string` i.e. always including the property, and it having either the value null or a real string. This was causing an exception in nuclide-rpc. I made Nuclide robust against the possibility in D8464870.

But we should still amend lsp_fmt to produce `title` in the conventional way.

Reviewed By: arxanas

Differential Revision: D8464931

fbshipit-source-id: 222127c2fa2917e6d1030a1c53f2b2a61ec6a795