[Easy] Small fix for --profile table
Summary:
When you run Flow with `--profile`, it prints out a small table with the wall & cpu times for each step. However, it's possible for some un-profiled code to take a long time. To highlight this, if there's ever a gap of 1% of total time between steps, we print out an <Unknown> step.

However, we were failing to print out an <Unknown> step if it came before the first step (after we start profiling but before the first timer). This fixes that lapse.

Reviewed By: fishythefish

Differential Revision: D8424831

fbshipit-source-id: 37a0b61b47464eb5b2180238977f089588b36083