Add unparsed and ordered_libs to env
Summary:
The shared memory heaps don't provide anyway to iterate through them. If you want to dump a heap to a file, you need keep a separate list of all the keys.

To grab all the data for parsed files, we can use env.files. But we actually also need to dump InfoHeap for all the unparsed files. So we need env.unparsed.

In addition, we want to avoid walking the file looking for libraries, so we need to save a list of the libraries. And since we need to merge them in the same order, we need this list to be ordered.

Reviewed By: avikchaudhuri

Differential Revision: D8425260

fbshipit-source-id: 7563d7de86735db14f35754da0f34558720df884