Errors.error mapper
Summary:
I have decided to include local errors in the saved state. This means I need to normalize all the locations in an `Errors.error`. The easiest way to do this and make sure it stays correct is to add a mapper.

I did a similar thing for `File_sig.t` in {D8408654}

Reviewed By: nmote

Differential Revision: D8472281

fbshipit-source-id: 533df4eee4936df1b80aa50a938554bc6f885789