`flow save-state` command
Summary:
`flow save-state --root path/to/root --out foo.saved_state` will tell the Flow server running for the "path/to/root" root to save its state in "foo.saved_state".

The saved state file will include all the information needed to skip the parsing step of initialization. The information in the saved state file should also be normalized so it can be used by different checkouts of the same repository.

I have identified the following heaps as needing to be saved:

* PackageHeap (For `package.json` files, it maps filename to data extracted from the `package.json`)
* ReversePackageHeap (Maps module names to the directory containing the `package.json`)
* NameHeap (Maps module name to filename)
* InfoHeap (Maps filename to info about the module)
* ResolvedRequiresHeap (Maps filename to info about all its imports and what they resolve to)
* FileSigHeap (Maps filename to a big, complicated data structure)

And actually the `NameHeap` and `ReversePackageHeap` can be recovered from the `InfoHeap` and `PackageHeap`, so we don't save those.

At the moment, normalization just means making all absolute paths relative and stripping out the builtin flowlibs

Reviewed By: avikchaudhuri

Differential Revision: D8432235

fbshipit-source-id: 6efe8ee42f8623da636246d7e93f18bd4e41278c