Add ref kind to each reference returned by the server-side find-refs function
Summary:
This doesn't affect the client output or even what is sent across the wire. It
simply makes it so that other callers of the find-refs function have this
information.

Right now I only have the `Other` kind for the sake of splitting diffs. I'll add
actual information next. For starters, we want to be able to differentiate
between references to a local variable and references to object properties.

It would not be sufficient to return one kind for an entire set of references.
We need this information on each reference because a single find-refs query can
actually return multiple kinds of references. For example, global find-refs on a
named ES6 export is treated as object find-refs, since consumers can get an
object representing the entire module via `import *`. However, in special cases
we also find local uses of the imported symbols. So, object propery references
and local variable references can intermingle. See the tests for this case for a
more concrete example.

Reviewed By: panagosg7

Differential Revision: D8452895

fbshipit-source-id: d41544706f593abdd00d8b12eaddad61cf7465ac